refactor(ActionList): move out shared values into module
