+++ src/ActionList/LinkItem.tsx
@@ -2,7 +2,8 @@ import React from 'react'
 import {ForwardRefComponent as PolymorphicForwardRefComponent} from '../utils/polymorphic'
 import Link from '../Link'
 import {SxProp, merge} from '../sx'
-import {Item, ActionListItemProps} from './Item'
+import {Item} from './Item'
+import {ActionListItemProps} from './shared'
 
 // adopted from React.AnchorHTMLAttributes
 type LinkProps = {
