+++ src/ActionList/Visuals.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Box from '../Box'
 import {SxProp, merge} from '../sx'
 import {get} from '../constants'
-import {getVariantStyles, Slot, ItemContext, TEXT_ROW_HEIGHT} from './Item'
+import {getVariantStyles, Slot, ItemContext, TEXT_ROW_HEIGHT} from './shared'
 
 type VisualProps = SxProp & React.HTMLAttributes<HTMLSpanElement>
