+++ src/ActionList/index.ts
@@ -8,7 +8,7 @@ import {LeadingVisual, TrailingVisual} from './Visuals'
 
 export type {ActionListProps} from './List'
 export type {ActionListGroupProps} from './Group'
-export type {ActionListItemProps} from './Item'
+export type {ActionListItemProps} from './shared'
 export type {ActionListLinkItemProps} from './LinkItem'
 export type {ActionListDividerProps} from './Divider'
 export type {ActionListDescriptionProps} from './Description'
